--- types/tabs/tabs.ts
@@ -1,9 +1,0 @@
-export interface Tab {
-    id: string;
-    content: React.ReactNode;
-}
-
-export interface TabsProps {
-    tabs?: Tab[];
-    className?: string;
-}
